Finished 99% of CSS and 1/2 of HTML comments
Mostly finished the CSS styles and half of the comments for HTML. Still need to finish the other half of HTML comments, Comments for CSS, and the last style for text area boxes. For the CSS, I worked mostly with changing widths, floating elements left, and margins. So far, I effected the fieldsets and their legends, selection lists, text area boxes, et cetera.